Remove setup_hooks.dart and move _setupHooks back into natives.dart
_setupHooks had been extracted into a separate file as part of Impeller Scene, which was later removed from the engine.

setup_hooks.dart was not distributed with Flutter releases, and the reference to a nonexistent file in the dart:ui library caused issues for tools like source_gen
