+++ engine/src/flutter/lib/ui/ui.dart
@@ -38,6 +38,5 @@ part 'platform_isolate.dart';
 part 'plugins.dart';
 part 'pointer.dart';
 part 'semantics.dart';
-part 'setup_hooks.dart';
 part 'text.dart';
 part 'window.dart';
